backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
alembic
python-dotenv
passlib
python-jose
python-dotenv
